Reorganize a little bit
