+++ crates/notebook/src/jupyter/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod message;
+pub(crate) mod python;
